"bedita cleanup" shell script doesn't remove file named "empty"
"empty" files are used to track empty folder in git and them must not be removed in clenup operations

+++ bedita-app/vendors/shells/bedita_base.php
@@ -181,9 +181,11 @@ class BeditaBaseShell extends Shell {
 		}
         foreach ($list[1] as $f) {
         	$file = new File($folder->path.DS.$f);
-        	if(!$file->delete()) {
-                throw new Exception("Error deleting file $f");
-            }
+        	if ($file->name != "empty") {
+	        	if(!$file->delete()) {
+	                throw new Exception("Error deleting file $f");
+	            }
+	        }
         }
         return ;
     }
